@@ -2,17 +2,10 @@ package com.github.jonasxpx;
 
 import com.github.jonasxpx.facade.FaturaReader;
 
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    
+public class App {
 
-    public static void main( String[] args ) throws Exception
-    {
-       new FaturaReader()
-        .start();
+    public static void main(String[] args) throws Exception {
+        new FaturaReader()
+                .read();
     }
 }
