@@ -6,28 +6,25 @@ import java.util.Iterator;
 import org.apache.poi.ss.usermodel.Sheet;
 import org.apache.poi.ss.usermodel.Workbook;
 import org.apache.poi.ss.usermodel.WorkbookFactory;
+
 import static java.lang.String.format;
 
 
-public class FaturaReader extends Thread {
+public class FaturaReader {
 
-    @Override
-    public void run() {
-         try (Workbook workbook = WorkbookFactory.create(new File("extrato.xls"))){
-            
-            Integer cellCount = workbook.getNumCellStyles();
-            Integer sheetsCount = workbook.getNumberOfSheets();
+   public void read() {
+       try (Workbook workbook = WorkbookFactory.create(new File("extrato.xls"))) {
+           Integer cellCount = workbook.getNumCellStyles();
+           Integer sheetsCount = workbook.getNumberOfSheets();
 
-            Iterator<Sheet> sheetIterator = workbook.sheetIterator();
-            
-            while(sheetIterator.hasNext()) {
-                Sheet sheet = sheetIterator.next();
-                System.out.println(format("Sheet name: %s", sheet.getSheetName()));
-            }
-            
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-    }
+           Iterator<Sheet> sheetIterator = workbook.sheetIterator();
 
+           while (sheetIterator.hasNext()) {
+               Sheet sheet = sheetIterator.next();
+               System.out.printf("Sheet name: %s%n", sheet.getSheetName());
+           }
+       } catch (Exception e) {
+           e.printStackTrace();
+       }
+   }
 }
